requirements.txt
fastapi==0.109.0
uvicorn==0.27.0
langchain==0.1.0
qdrant-client==1.7.3
psycopg2-binary==2.9.9
redis==5.0.1
pydantic==2.5.3
celery==5.3.6
python-multipart==0.0.6
pypdf==4.0.1
python-docx==1.1.0
openai==1.12.0
requests==2.31.0
